Python/ETL_With_Gcp/requirements.txt
pandas
google-cloud-storage
boto3